+++ src/main/java/cs545_project/online_market/service/AddressService.java
@@ -5,7 +5,7 @@ import cs545_project.online_market.domain.Address;
 import cs545_project.online_market.domain.User;
 
 public interface AddressService {
-    public Address findById(int id);
+    public Address findById(long id);
     public void delete(Address address);
     public Address save(Address address);
     public Address createOrUpdate(AddressRequest addressRequest, String type, User user);
